Create 1498. Number of Subsequences That Satisfy the Given Sum Condition
approach is simple, we are using two pointer approach on sorted array.
why we are doing sorting?? because the question say that we need min and max element sum for every subsequence which should be <= target value. So order doesn't matter here.
now we are checking here that, if left and right value sum <= target, 
then by taking that min and max value we can make 2^(right-left) subsequence(formula in case we didn't know that). because we only need to take care about max and min element and rest other element doesn't matter if they are in subsequence or not, answer will always be less or equals to target.

Thank You!!!